Move bookstore demo database config to app db
