+++ demos/bookstore/app/db.ts
@@ -1,73 +1,25 @@
-import * as fs from 'node:fs'
-import * as os from 'node:os'
 import * as path from 'node:path'
-import { DatabaseSync } from 'node:sqlite'
-import { fileURLToPath } from 'node:url'
-import { createDatabase } from 'remix/data-table'
-import { createMigrationRunner } from 'remix/data-table/migrations'
+import { createDatabase, createMigrator } from 'remix/data-table'
 import { loadMigrations } from 'remix/data-table/migrations/node'
 import { createSqliteDatabaseAdapter } from 'remix/data-table/sqlite'
 
-import { books, orderItems, orders, users } from './schema.ts'
-import { hashPassword } from '../utils/password-hash.ts'
+import { books, orderItems, orders, users } from './data/schema.ts'
+import { hashPassword } from './utils/password-hash.ts'
 
-let databaseFilePath: string
-let testDatabaseDirectoryPath: string | undefined
+export const db = createDatabase(
+  createSqliteDatabaseAdapter({
+    filename:
+      process.env.NODE_ENV === 'test'
+        ? ':memory:'
+        : path.join(import.meta.dirname, '../db/bookstore.sqlite'),
+  }),
+)
 
-if (process.env.NODE_ENV === 'test') {
-  testDatabaseDirectoryPath = fs.mkdtempSync(path.join(os.tmpdir(), 'remix-bookstore-'))
-  databaseFilePath = path.join(testDatabaseDirectoryPath, 'bookstore.sqlite')
-} else {
-  let databaseDirectoryUrl = new URL('../../db/', import.meta.url)
-  databaseFilePath = fileURLToPath(new URL('bookstore.sqlite', databaseDirectoryUrl))
+export const migrator = createMigrator(
+  await loadMigrations(path.join(import.meta.dirname, '../db/migrations')),
+)
 
-  fs.mkdirSync(fileURLToPath(databaseDirectoryUrl), { recursive: true })
-}
-
-const migrationsDirectoryPath = fileURLToPath(new URL('../../db/migrations/', import.meta.url))
-
-const sqlite = new DatabaseSync(databaseFilePath)
-sqlite.exec('PRAGMA foreign_keys = ON')
-const adapter = createSqliteDatabaseAdapter(sqlite)
-
-export const db = createDatabase(adapter)
-
-let initializePromise: Promise<void> | null = null
-
-export async function initializeBookstoreDatabase(): Promise<void> {
-  if (!initializePromise) {
-    initializePromise = initialize()
-  }
-
-  await initializePromise
-}
-
-export function closeBookstoreDatabase(): void {
-  if (process.env.NODE_ENV === 'test' && process.platform === 'win32') {
-    // DatabaseSync.close() can crash during Windows test process shutdown.
-    // Each test file runs in its own process, and the runner discards temp files.
-    return
-  }
-
-  if (sqlite.isOpen) {
-    sqlite.close()
-  }
-
-  if (testDatabaseDirectoryPath) {
-    fs.rmSync(testDatabaseDirectoryPath, { recursive: true, force: true })
-    testDatabaseDirectoryPath = undefined
-  }
-}
-
-if (process.env.NODE_ENV === 'test') {
-  process.once('exit', closeBookstoreDatabase)
-}
-
-async function initialize(): Promise<void> {
-  let migrations = await loadMigrations(migrationsDirectoryPath)
-  let migrationRunner = createMigrationRunner(adapter, migrations)
-  await migrationRunner.up()
-
+export async function seed(): Promise<void> {
   let booksCount = await db.count(books)
   if (booksCount === 0) {
     await db.createMany(books, [
@@ -205,3 +157,4 @@ async function initialize(): Promise<void> {
     ])
   }
 }
+
